requirements.txt
PyPDF2
python-docx
shutil
pathlib